+++ data-raw/upgrade.R
@@ -1,41 +1,3 @@
-library(magrittr)
-
-html <- xml2::read_html("https://www.sqlite.org/download.html")
-
-latest_name <- html %>%
-  rvest::html_nodes("a") %>%
-  xml2::xml_find_all('//*[(@id = "a1")] | //*[(@id = "a2")]') %>%
-  rvest::html_text() %>%
-  grep("amalgamation", ., value = TRUE) %>%
-  .[1]
-
-tmp <- tempfile()
-
-year <- as.integer(strftime(Sys.Date(), "%Y"))
-latest <- paste0("https://sqlite.org/", year, "/", latest_name)
-
-tryCatch(
-  download.file(latest, tmp),
-  warning = function(e) {
-    if (grepl("404", conditionMessage(e))) {
-      latest <- paste0("https://sqlite.org/", year - 1, "/", latest_name)
-      download.file(latest, tmp)
-    }
-  }
-)
-
-unzip(tmp, exdir = "src/vendor/sqlite3", junkpaths = TRUE)
-unlink("src/vendor/sqlite3/shell.c")
-file.rename("src/vendor/sqlite3/sqlite3ext.h", "src/vendor/extensions/sqlite3ext.h")
-
-
-tmp_source_zip <- tempfile()
-latest_code <- "https://www.sqlite.org/src/zip/sqlite.zip?r=release"
-download.file(latest_code, tmp_source_zip)
-
-tmp_source_dir <- tempdir()
-unzip(tmp_source_zip, exdir = tmp_source_dir)
-
 
 # Regular expression source code
 register_misc_extension <- function(name) {
@@ -43,7 +5,7 @@ register_misc_extension <- function(name) {
   if (!dir.exists(ext_dir))
     dir.create(ext_dir, recursive = TRUE)
 
-  text <- readLines(paste0(tmp_source_dir, "/sqlite/ext/misc/", name, ".c"))
+  text <- readLines(paste0(tmp_source_dir, "/ext/misc/", name, ".c"))
   text <- gsub(" +$", "", text)
   writeLines(text, paste0(ext_dir, name, ".c"))
 
@@ -59,77 +21,81 @@ register_misc_extension <- function(name) {
   writeLines(lines, paste0("src/ext-", name, ".c"))
 }
 
+
+
+# Identify latest version and download source tar from Github
+html <- xml2::as_list(
+  xml2::read_html("https://github.com/sqlcipher/sqlcipher/releases/latest"))[[1]]
+ver <- regmatches(html$head$title[[1]],
+                  regexpr("[0-9]+\\.[0-9]+\\.[0-9]+",
+                          html$head$title[[1]]))
+
+tmp_source_tar <- tempfile()
+latest_code <- paste0("https://github.com/sqlcipher/sqlcipher/archive/refs/tags/v", ver, ".tar.gz")
+download.file(latest_code, tmp_source_tar)
+
+# Extract and make 'amalgamation' file
+tmp_source_dir <- tempdir()
+untar(tmp_source_tar, exdir = tmp_source_dir)
+
+tmp_source_dir <- paste0(tmp_source_dir, "/sqlcipher-", ver)
+
+system(paste0("cd ", tmp_source_dir, " && ./configure --enable-all && make sqlite3.c"))
+
+# Copy main sources and headers
+text <- readLines(paste0(tmp_source_dir, "/sqlite3.h"))
+text <- gsub(" +$", "", text)
+writeLines(text, paste0("src/vendor/sqlite3/sqlite3.h"))
+tmp_sqlite_ver <- grepv("#define SQLITE_VERSION ", text)
+sqlite_ver <- regmatches(tmp_sqlite_ver, regexpr("[0-9]+\\.[0-9]+\\.[0-9]+",tmp_sqlite_ver))
+rm(tmp_sqlite_ver)
+
+text <- readLines(paste0(tmp_source_dir, "/sqlite3.c"))
+text <- gsub(" +$", "", text)
+writeLines(text, paste0("src/vendor/sqlite3/sqlite3.c"))
+
+# Copy extensions sources and headers
+text <- readLines(paste0(tmp_source_dir, "/src/sqlite3ext.h"))
+text <- gsub(" +$", "", text)
+writeLines(text, paste0("src/vendor/extensions/sqlite3ext.h"))
+rm(text)
+
 register_misc_extension("regexp")
 register_misc_extension("series")
 register_misc_extension("csv")
 register_misc_extension("uuid")
 
-
+# Commit changes
 if (any(grepl("^src/", gert::git_status()$file))) {
-  branch <- paste0("f-", sub("[.][^.]*$", "", latest_name))
-  message("Changes detected, creating branch: ", branch)
-
-  version <- sub("^.*-([0-9])([0-9][0-9])(?:0([0-9])|([1-9][0-9]))[0-9]+[.].*$", "\\1.\\2.\\3\\4", latest_name)
-
-  old_branch <- gert::git_branch()
-  message("Old branch: ", old_branch)
-
-  gert::git_branch_create(branch)
+  # branch <- paste0("dev_bundled_source-", ver)
+  # message("Changes detected, creating branch: ", branch)
+  #
+  # old_branch <- gert::git_branch()
+  # message("Old branch: ", old_branch)
+  #
+  # gert::git_branch_create(branch)
   gert::git_add("src")
 
-  commit_msg <- paste0("Upgrade bundled SQLite to ", version)
+  commit_msg <- paste0("build: upgrade bundled SQLCipher to ", ver, " (SQLite ", sqlite_ver, ")")
   message("Commit message: ", commit_msg)
   gert::git_commit(commit_msg)
 
-  title <- paste0("feat: ", commit_msg)
-
-  # Force-pushing: this job is run daily, will give a daily notification
-  # and still succeed
-  message("Pushing branch")
-  gert::git_push(force = TRUE)
-
-  message("Checking if PR exists")
-  existing_pr <- gh::gh(
-    "/repos/r-dbi/RSQLite/pulls",
-    head = paste0("r-dbi:", branch), base = old_branch,
-    state = "open"
-  )
-
-  if (length(existing_pr) > 0) {
-    message("Nudging")
-    gh::gh(
-      paste0("/repos/r-dbi/RSQLite/issues/", existing_pr[[1]]$number, "/comments"),
-      body = "PR updated.",
-      .method = "POST"
-    )
-  } else {
-    message("Opening PR")
-    pr <- gh::gh(
-      "/repos/r-dbi/RSQLite/pulls",
-      head = branch, base = old_branch,
-      title = title, body = ".",
-      .method = "POST"
-    )
-
-    message("Tweaking PR body")
-    body <- paste0("NEWS entry will be picked up by fledge from the PR title.")
-
-    gh::gh(
-      paste0("/repos/r-dbi/RSQLite/pulls/", pr$number),
-      body = body,
-      .method = "PATCH"
-    )
-
-    message("Adding label")
-    gh::gh(
-      paste0("/repos/r-dbi/RSQLite/issues/", pr$number),
-      labels = c("mergequeue", "mergequeue"),
-      .method = "PATCH"
-    )
-
-    message("Bumping main branch to run CI/CD")
-    gert::git_branch_checkout(old_branch)
-    system2("git", c("commit", "-m", "'chore: Bump main branch for CI/CD'", "--allow-empty"))
-    gert::git_push(force = FALSE)
-  }
 }
+
+# Patch new sources and commit
+for (f in dir("patch", full.names = TRUE)) {
+  message("Applying ", f)
+  stopifnot(system(paste0("patch -p1 -i ", f)) == 0)
+}
+
+if (any(grepl("^src/", gert::git_status()$file))) {
+
+  gert::git_add("src")
+
+  commit_msg <- paste0("fix: apply patches")
+  message("Commit message: ", commit_msg)
+  gert::git_commit(commit_msg)
+}
+
+gert::git_branch_checkout(old_branch)
+gert::git_merge(branch)